[HeaderNav] Fix stories
