+++ assets/javascripts/kitten/components/organisms/header-nav/config.js
@@ -1,3 +1,1 @@
-export const MOBILE_HEADER_HEIGHT = 65
-export const TABLET_HEADER_HEIGHT = 80
-export const DESKTOP_HEADER_HEIGHT = 80
+export const HEADER_HEIGHT = 65
